Now calling get by id and update tests to match

+++ RecordShop/RecordShop.Api/Services/ArtistService.cs
@@ -87,6 +87,10 @@ namespace RecordShop.Api.Services
 
         public async Task<bool> DeleteArtistByIdAsync(int id)
         {
+            var artist = await _artistRepository.GetArtistByIdAsync(id);
+
+            if (artist == null) return false;
+
             return await _artistRepository.DeleteArtistByIdAsync(id);
         }
     }
